+++ foxes/input/states/__init__.py
@@ -17,6 +17,7 @@ from .field_data import WeibullField as WeibullField
 
 from .point_cloud_data import WeibullPointCloud as WeibullPointCloud
 from .point_cloud_data import PointCloudData as PointCloudData
+from .point_cloud_data import TurbinePointCloud as TurbinePointCloud
 
 from .multi_height import MultiHeightStates as MultiHeightStates
 from .multi_height import MultiHeightTimeseries as MultiHeightTimeseries